Working on all formats

+++ scrap.py
@@ -10,6 +10,8 @@ import time
 from bs4 import BeautifulSoup
 import os
 import requests
+from PyPDF2 import PdfMerger
+from selenium.webdriver.common.action_chains import ActionChains
 
 '''
 car_models = driver.find_elements_by_class_name("carbox") # return drop menu of each car item
@@ -20,27 +22,24 @@ pdf_links = driver.find_elements_by_xpath("//a[contains(@onclick, 'pdf')]") # ge
 
 '''
 
-from selenium.webdriver.common.action_chains import ActionChains
-outfolder = "pdfs/"
-language = "en_us"
-session = requests.Session()
-
-
 def merge_pdfs(pdf_list):
     '''
     Merger all the pdfs in list 
     '''
+    output_path = pdf_list[0].replace(pdf_list[0].split('/')[-1],"merged")+os.path.splitext(pdf_list[0])[1]
 
-    pdf_merger = PdfFileMerger()
+    if not os.path.exists(output_path):
 
-    # Add PDFs to the merger
-    for pdf_path in pdf_paths:
-        with open(pdf_path, 'rb') as pdf_file:
-            pdf_merger.append(pdf_file)
+        pdf_merger = PdfMerger()
 
-    # Write the merged PDF to the output file
-    with open(output_path, 'wb') as output_file:
-        pdf_merger.write(output_file)
+        for pdf_path in pdf_list:
+            with open(pdf_path, 'rb') as pdf_file:
+                pdf_merger.append(pdf_file)
+
+        # Write the merged PDF to the output file
+        print("Mering and Saving {}".format(output_path))
+        with open(output_path, 'wb') as output_file:
+            pdf_merger.write(output_file)
 
 def refresh_page():
 
@@ -85,6 +84,14 @@ def extract_car_periods(html):
 
 
 def download_file(url, savepath):
+    #Checking is folder exists or not
+    save_folder = savepath.replace(savepath.split('/')[-1], "")
+
+    if not os.path.exists(save_folder):
+        print("Creating {}".format(save_folder))
+        os.makedirs(save_folder)
+
+
     if not os.path.exists(savepath):
         print("Saving {} into {}".format(url, savepath))
         response = session.get(url)
@@ -92,35 +99,17 @@ def download_file(url, savepath):
             f.write(response.content)
 
 
-def download(beautiful_perioids, selenimum_periods,index ):
+def download(beautiful_perioids, selenimum_periods,period, index):
     pdf_info = beautiful_perioids[index]
 
-
-    outpath = outfolder+pdf_info[2]+"/"
-
-    
-
-    if not os.path.exists(outpath):
-        os.makedirs(outpath)
-
+    outpath = outfolder+pdf_info[1]+"-"+pdf_info[2]+"/"+period+"/"
 
 
     if pdf_info[0]=="pdfsimple":#latest single pfds
         pdf_url = url+"modeles/"+pdf_info[1]+"/"+pdf_info[2]+"/"+pdf_info[3]+"/"+language+"/"+ pdf_info[1]+"_"+pdf_info[2]+"_"+pdf_info[3]+"_"+language+".pdf"
-
         pdf_file = pdf_url.split('/')[-1]
-
         savefile = outpath+pdf_file
-
-        if not os.path.exists(savefile):
-            print("Downloading: ",pdf_url)
-
-
-            response = session.get(pdf_url)
-
-            # Save the PDF file to disk
-            with open(savefile, 'wb') as f:
-                f.write(response.content)
+        download_file(pdf_url, savefile)
 
 
     elif pdf_info[0]=="pdf": #Multiple pages pdfs
@@ -130,16 +119,10 @@ def download(beautiful_perioids, selenimum_periods,index ):
         soup = BeautifulSoup(driver.page_source, 'html.parser')
  
         image_url = soup.find("div", id="doczone").find('img').attrs['src']
-
-
         
-        img_path = outfolder+car_brand+img_folder+os.path.join(os.path.splitext(os.path.basename(image_url))[0], image_url.split('/')[-1])
+        img_path = outpath+"FrontCovers/"+ image_url.split('/')[-1]
 
-        if not os.path.exists(os.path.split(img_path)[0]):
-            os.makedirs(os.path.split(img_path)[0])
-
-
-        download_file(image_url, img_path)
+        download_file(url+image_url, img_path)
 
 
         pages = soup.find_all("a", class_="prglinkpdf")
@@ -151,31 +134,22 @@ def download(beautiful_perioids, selenimum_periods,index ):
                 pdf_links.append(href)
 
 
-        outpath = outfolder+car_brand+pdf_info[1]+"-"+pdf_info[2]+"/"+pdf_info[3]+"/"
-
-        if not os.path.exists(outpath):
-            print("Creating dir: ", outpath)
-            os.makedirs(outpath)
-
         pdf_list = []
         for pdf_link in pdf_links:
-            savepath = outfolder+car_brand+pdf_info[1]+"-"+pdf_info[2]+"/"+pdf_info[3]+"/"+pdf_link
+            savepath = outpath+pdf_link
             source_path = url+"modeles/"+pdf_info[1]+"/"+pdf_info[2]+"/"+pdf_info[3]+"/"+language+"/"+ pdf_link
 
             pdf_list.append(savepath)
 
             download_file(source_path, savepath)
 
+        merge_pdfs(pdf_list)
+
         time.sleep(2)
         refresh_page()
 
 
-    elif pdf_info[0]=="eGuide":
-        outpath = outfolder+car_brand+pdf_info[1]+"-"+pdf_info[2]+"/"+pdf_info[3]+"/"
-        if not os.path.exists(outpath):
-            os.makedirs(outpath)
-
-
+    elif pdf_info[0]=="eGuide":   
         selenimum_periods[i].click()
         time.sleep(2)
         soup_new = BeautifulSoup(driver.page_source, 'html.parser')
@@ -203,9 +177,10 @@ def download(beautiful_perioids, selenimum_periods,index ):
         refresh_page()
 
 
-
-
 driver = webdriver.Chrome(ChromeDriverManager().install())
+language = "en_us"
+session = requests.Session()
+
 url = "https://service.citroen.com/ACddb/"
 
 img_folder  = "Front_covers/"
@@ -217,14 +192,14 @@ car_hovers = driver.find_elements_by_class_name("carbox") # returnls name of ite
 
 time.sleep(1)
 
-car_index = 6
+car_index = 4
 car_hover = car_hovers[car_index]
 
 print(car_hover.text)
+outfolder = "pdfs/"+car_brand+car_hover.text+"/"
+
 
 car_hover.click()
-
-
 
 models = car_hover.find_elements_by_class_name("cartype") # returnls name of items
 
@@ -247,26 +222,18 @@ else:
 
         if "dag" not  in models[i].get_attribute("title"):
             print("Fetching pdfs from {}".format(models[i].get_attribute("title")))
-
         
             models[i].click()
 
-
-
-            # Fetching periods of each model
-
-            # Find all <li> elements under the <ul> element
-
             beautiful_periods, date_periods = extract_car_periods(driver.page_source)
-
-
 
             for i in range(len(beautiful_periods)):
                 element = driver.find_element_by_id("licarperiodelist")
                 selenium_period_list = element.find_elements_by_tag_name("li")
 
                 print(beautiful_periods[i], "::::", date_periods[i])
-                download(beautiful_periods, selenium_period_list, i)
+
+                download(beautiful_periods, selenium_period_list,date_periods[i].replace("/","-"), i)
 
 
             #Refreshing the page again
@@ -277,8 +244,6 @@ else:
             car_hover = car_hovers[car_index]
             car_hover.click()
             models = car_hover.find_elements_by_class_name("cartype") # returnls name of items
-
-
 
     
 #Close the browser window
